+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.6.0] - ?
 ### Changed
 - Massive reformatting and clean up of the entire code base.
+- We now use Apache HTTP Client to make all of our HTTP requests.
 
 ### Added
 - Added additional documentation.